+++ attributes/ngx_pagespeed.rb
@@ -1,13 +1,13 @@
 # # # # # # # # #
 # ngx_pagespeed #
 # # # # # # # # #
-default['nginx']['ngx_pagespeed']['version'] = "1.8.31.2-beta"
+default['nginx']['ngx_pagespeed']['version'] = "1.8.31.4-beta"
 default['nginx']['ngx_pagespeed']['url'] = "https://codeload.github.com/pagespeed/ngx_pagespeed/tar.gz/v#{node['nginx']['ngx_pagespeed']['version']}"
-default['nginx']['ngx_pagespeed']['checksum'] = "8cd1bca463b4810cf2e7aa738e7a9fe4c40075ca9bf9a8e14e1a470a9e0540c589a7ee0a0a12cda1e479c7fbe9ea2f6e080ec4333ccabef6dd335691c88188e2"
+default['nginx']['ngx_pagespeed']['checksum'] = "41417cf54321f3fd80e9529808b58cbefb2b51977fc41ef1ed048fe4993fe44f5d3bbf60cfc4c44bd19935019829012681ee272e5d48ead3f6953c1a222a69c7"
 
-default['nginx']['ngx_pagespeed']['psol']['version'] = "1.8.31.2"
+default['nginx']['ngx_pagespeed']['psol']['version'] = "1.8.31.4"
 default['nginx']['ngx_pagespeed']['psol']['url'] = "https://dl.google.com/dl/page-speed/psol/#{node['nginx']['ngx_pagespeed']['psol']['version']}.tar.gz"
-default['nginx']['ngx_pagespeed']['checksum'] = "e68c7bc8f5306f249ca7bbaad3d8f15bdcbda5e010821dce21ca0059e128fe83f3d8bd6f8a7e606169e51d65ef29f549c81636f84783b9c721092a16c7afac38"
+default['nginx']['ngx_pagespeed']['checksum'] = "90866d2b7c0397ea6b216ca1b44b43a80a6d5045b9cc9f7eca1c03d42418e8d143ee1396de7da7fa61b86845947112ba9183ad864f33cfe5523e7ea43b3aa932"
 
 default['nginx']['ngx_pagespeed']['cache-url'] = "/var/ngx_pagespeed_cache"
 default['nginx']['ngx_pagespeed']['cache-size'] = "512M"
@@ -38,5 +38,6 @@ default['nginx']['ngx_pagespeed']['FileCacheInodeLimit'] = 500000
 default['nginx']['ngx_pagespeed']['LRUCacheKbPerProcess'] = 8192
 default['nginx']['ngx_pagespeed']['LRUCacheByteLimit'] = 16384
 default['nginx']['ngx_pagespeed']['CacheFlushPollIntervalSec'] = 600000
+default['nginx']['ngx_pagespeed']['DownstreamCachePurgeMethod'] = "PURGE"
 default['nginx']['ngx_pagespeed']['DownstreamCacheRewrittenPercentageThreshold'] = 95
 default['nginx']['ngx_pagespeed']['RetainComment'] = "esi*"