+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This file is used to list changes made in each version of dop_nginx.
 ## 1.0.9  (July 2, 2014)
 
 * add dependency to python and install [ngxtop](https://github.com/lebinh/ngxtop)
+* upgrade ngx_pagespeed
+* add ngx_pagespeed configs
 
 ## 1.0.8  (May 24, 2014)